Add initial foundation for cromwell in analysis runner
